recom_service/requirements.txt
flask
pandas
numpy
scikit-learn
scipy
scikit-surprise
